# src/multi-chat/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kuwa API",
        "description": "API definition for KuwaClient service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/create/base_model": {
            "post": {
                "tags": [
                    "Models"
                ],
                "summary": "Create a base model",
                "operationId": "c10fabc41138fda505a0a8e8a5210556",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BaseModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Model created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/read/models": {
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "List base models",
                "operationId": "432ffdc597ac392760cd65b5f1bc88c1",
                "responses": {
                    "200": {
                        "description": "List of models"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BaseModel": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "access_code": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "other_field": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}